+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/responseprocessors/CacheV2ResponseProcessor.java
@@ -118,11 +118,12 @@ public class CacheV2ResponseProcessor implements ResponseProcessor {
             } catch (Exception e) {
                 //mark and log the cache put failure
                 CACHE_SET_FAILURES.mark(1);
-                assert valueString != null;
                 RequestLog.record(new BardCacheInfo(
                         cacheKey,
+                        cacheKey.length(),
                         getMD5Cksum(cacheKey),
-                        valueString.length()));
+                        valueString != null ? valueString.length() : 0
+                ));
                 LOG.warn(
                         "Unable to cache {}value of size: {}",
                         valueString == null ? "null " : "",