+++ CHANGELOG.md
@@ -74,7 +74,7 @@ Current
     * Created `LegacyGenerator` as a bridge interface from the existing constructor based api request impls and the factory based value object usage.
 
 ### Added:
-- [Add logging for cache put failures](https://github.com/yahoo/fili/pull/??)
+- [Add logging for cache put failures](https://github.com/yahoo/fili/pull/1114)
    * Added `BardCacheInfo` logging for cache put failures.
    
 - [Add withLogicalMetricInfo to MetricInstance class](https://github.com/yahoo/fili/pull/1105)